maven.yml: add `-B` to the maven command line
Add the `-B` (batch mode) to the `mvn` command line. In interactive mode maven prints many lines of download progress output filling the console output with thousands of useless lines. Ideally the github-actions console would interpret `\r' as carriage return (go back to the beginning of the current line and overwrite it). As long as that is not the case using `-B` is a good workaround.