Default to publishing with provenance in npm publish workflow
Update the `npm-publish.yml` starter workflow to publish with provenance to encourage users publishing to npm to use provenance.

A prerequisite for this is upgrading Node.js to v20 because it ships with an npm version that supports the `--provenance` flag. An alternative to this would be to add a `-run: npm install -g npm` step before publishing.