Rename README.md to Efikcoinstaking.sol
Efikcoin is a digital currency that is associated with the Efik people, an ethnic group in southeastern Nigeria. It is designed to promote economic development and financial inclusion within the Efik community and beyond. The currency aims to facilitate transactions, support local businesses, and encourage investment in the region. Like other cryptocurrencies, Efikcoin operates on blockchain technology, which ensures secure and transparent transactions.💯🚀